chg: [misp-galaxy] removed unused derivated_from field from MISP galaxy schema
This has been validated by @Delta-Sierra and schema got updated.

@adulau updated the Internet-Draft following the approval of the TC
at unanimity following a debate of 7 seconds which concluded in
the approval of the removal of the field. The user-groups relying
on the galaxy schema approved that the field was not used and was
superseded by the 'related' field which is more convenient and
already implemented by @cvandeplas (ref:
https://github.com/MISP/misp-galaxy/blob/master/tools/gen_mapping.py)

As the TC strictly follows the PMF methodology, the commit was signed off
by the quorum required to validate a TC decision.

Ref: https://twitter.com/Iglocska/status/1039214679267201026 (TC
approval)